chatbot/pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["bipin <bipin.mishra@atlascopco.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "3.8.5"
aiosignal = "1.3.1"
asgiref = "3.7.2"
async-timeout = "4.0.3"
attrs = "23.1.0"
build = "0.10.0"
cachecontrol = "0.13.1"
cachetools = "5.3.1"
certifi = "2023.7.22"
cffi = "1.16.0"
charset-normalizer = "3.3.0"
cleo = "2.0.1"
crashtest = "0.4.1"
cryptography = "41.0.4"
distlib = "0.3.7"
django = "4.2.5"
dulwich = "0.21.6"
filelock = "3.12.4"
frozenlist = "1.4.0"
google-ai-generativelanguage = "0.3.2"
google-api-core = "2.12.0"
google-auth = "2.23.2"
google-generativeai = "0.2.0"
googleapis-common-protos = "1.60.0"
grpcio = "1.59.0"
grpcio-status = "1.59.0"
idna = "3.4"
importlib-metadata = "6.8.0"
installer = "0.7.0"
jaraco-classes = "3.3.0"
jeepney = "0.8.0"
jsonschema = "4.17.3"
keyring = "24.2.0"
more-itertools = "10.1.0"
msgpack = "1.0.7"
multidict = "6.0.4"
openai = "0.28.1"
packaging = "23.1"
pexpect = "4.8.0"
pkginfo = "1.9.6"
platformdirs = "3.10.0"
poetry = "1.6.1"
poetry-core = "1.7.0"
poetry-plugin-export = "1.5.0"
proto-plus = "1.22.3"
protobuf = "4.24.3"
ptyprocess = "0.7.0"
pyasn1 = "0.5.0"
pyasn1-modules = "0.3.0"
pycparser = "2.21"
pyproject-hooks = "1.0.0"
pyrsistent = "0.19.3"
python-dotenv = "1.0.0"
rapidfuzz = "2.15.1"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
rsa = "4.9"
secretstorage = "3.3.3"
shellingham = "1.5.3"
sqlparse = "0.4.4"
tomli = "2.0.1"
tomlkit = "0.12.1"
tqdm = "4.66.1"
trove-classifiers = "2023.9.19"
typing-extensions = "4.8.0"
urllib3 = "2.0.5"
virtualenv = "20.24.5"
yarl = "1.9.2"
zipp = "3.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
